+++ packages/checkbox-group/src/LionCheckboxGroup.js
@@ -1,12 +1,13 @@
+import { LitElement } from '@lion/core';
 import { ChoiceGroupMixin } from '@lion/choice-input';
-import { LionFieldset } from '@lion/fieldset';
+import { FormGroupMixin } from '@lion/fieldset';
 
 /**
  * A wrapper around multiple checkboxes
  *
  * @extends {LionFieldset}
  */
-export class LionCheckboxGroup extends ChoiceGroupMixin(LionFieldset) {
+export class LionCheckboxGroup extends ChoiceGroupMixin(FormGroupMixin(LitElement)) {
   constructor() {
     super();
     this.multipleChoice = true;
