+++ packages/fieldset/index.js
@@ -1,1 +1,2 @@
 export { LionFieldset } from './src/LionFieldset.js';
+export { FormGroupMixin } from './src/FormGroupMixin.js';
