+++ packages/fieldset/src/LionFieldset.js
@@ -1,499 +1,48 @@
-import { html, LitElement, SlotMixin } from '@lion/core';
-import { DisabledMixin } from '@lion/core/src/DisabledMixin.js';
-import { FormControlMixin, FormRegistrarMixin } from '@lion/field';
-import { getAriaElementsInRightDomOrder } from '@lion/field/src/utils/getAriaElementsInRightDomOrder.js';
-import { ValidateMixin } from '@lion/validate';
-import { FormElementsHaveNoError } from './FormElementsHaveNoError.js';
+import { LitElement } from '@lion/core';
+import { FormGroupMixin } from './FormGroupMixin.js';
 
 /**
- * LionFieldset: fieldset wrapper providing extra features and integration with lion-field elements.
+ * @desc LionFieldset is basically a 'sub form' and can have its own nested sub forms.
+ * It mimics the native <fieldset> element in this sense, but has all the functionality of
+ * a FormControl (advanced styling, validation, interaction states etc.). Also see
+ * FormGroup.
+ *
+ * LionFieldset enables the '_isFormOrFieldset' flag in FormRegistrarMixin. This makes .formElements
+ * act not only as an array, but also as an object (see FormRegistarMixin for more information).
+ * As a bonus, It can also group children having names ending with '[]'.
+ * This will be both helpful for forms and sub forms, which can contain sub forms as children
+ * as well and allow for a nested form structure, where other form groups
+ * (choice groups like radio-group, checkbox-group and (multi)select) don't:
+ * they should be considered 'end nodes' or 'leaves' of the form.
  *
  * @customElement lion-fieldset
  * @extends {LitElement}
  */
-export class LionFieldset extends FormRegistrarMixin(
-  FormControlMixin(ValidateMixin(DisabledMixin(SlotMixin(LitElement)))),
-) {
-  static get properties() {
-    return {
-      name: {
-        type: String,
-      },
-      submitted: {
-        type: Boolean,
-        reflect: true,
-      },
-      focused: {
-        type: Boolean,
-        reflect: true,
-      },
-      dirty: {
-        type: Boolean,
-        reflect: true,
-      },
-      touched: {
-        type: Boolean,
-        reflect: true,
-      },
-    };
-  }
-
-  get touched() {
-    return this.__touched;
-  }
-
-  set touched(value) {
-    const oldVal = this.__touched;
-    this.__touched = value;
-    this.requestUpdate('touched', oldVal);
-  }
-
-  get _inputNode() {
-    return this;
-  }
-
-  get modelValue() {
-    return this._getFromAllFormElements('modelValue');
-  }
-
-  set modelValue(values) {
-    this._setValueMapForAllFormElements('modelValue', values);
-  }
-
-  get serializedValue() {
-    return this._getFromAllFormElements('serializedValue');
-  }
-
-  set serializedValue(values) {
-    this._setValueMapForAllFormElements('serializedValue', values);
-  }
-
-  get formattedValue() {
-    return this._getFromAllFormElements('formattedValue');
-  }
-
-  set formattedValue(values) {
-    this._setValueMapForAllFormElements('formattedValue', values);
-  }
-
-  get prefilled() {
-    return this._everyFormElementHas('prefilled');
-  }
-
-  // TODO: can be protected? (or default and $form.$control('name') is api)
-  get formElementsArray() {
-    console.log('formElementsArray');
-    return Object.keys(this.formElements).reduce((result, name) => {
-      const element = this.formElements[name];
-      return result.concat(Array.isArray(element) ? element : [element]);
-    }, []);
-  }
-
-  // get formElements() {
-  //   return this.__formElements;
-  // }
-
-  // set formElements(v) {
-  //   this.__formElements = v;
-  // }
-
-  set fieldName(value) {
-    this.__fieldName = value;
-  }
-
-  get fieldName() {
-    const label =
-      this.label ||
-      (this.querySelector('[slot=label]') && this.querySelector('[slot=label]').textContent);
-    return this.__fieldName || label || this.name;
-  }
-
+export class LionFieldset extends FormGroupMixin(LitElement) {
   constructor() {
     super();
-    this.disabled = false;
-    this.submitted = false;
-    this.dirty = false;
-    this.touched = false;
-    this.focused = false;
-    this.formElements = {};
-    this.__addedSubValidators = false;
-
-    this._checkForOutsideClick = this._checkForOutsideClick.bind(this);
-
-    this.addEventListener('focusin', this._syncFocused);
-    this.addEventListener('focusout', this._onFocusOut);
-    this.addEventListener('dirty-changed', this._syncDirty);
-    this.addEventListener('validate-performed', this.__validate);
-
-    this.defaultValidators = [new FormElementsHaveNoError()];
-  }
-
-  connectedCallback() {
-    // eslint-disable-next-line wc/guard-super-call
-    super.connectedCallback();
-    this._setRole();
-  }
-
-  disconnectedCallback() {
-    super.disconnectedCallback(); // eslint-disable-line wc/guard-super-call
-
-    if (this.__hasActiveOutsideClickHandling) {
-      document.removeEventListener('click', this._checkForOutsideClick);
-      this.__hasActiveOutsideClickHandling = false;
-    }
-  }
-
-  updated(changedProps) {
-    super.updated(changedProps);
-
-    if (changedProps.has('disabled')) {
-      if (this.disabled) {
-        this.__requestChildrenToBeDisabled();
-      } else {
-        this.__retractRequestChildrenToBeDisabled();
-      }
-    }
-
-    if (changedProps.has('focused')) {
-      if (this.focused === true) {
-        this.__setupOutsideClickHandling();
-      }
-    }
-  }
-
-  __setupOutsideClickHandling() {
-    if (!this.__hasActiveOutsideClickHandling) {
-      document.addEventListener('click', this._checkForOutsideClick);
-      this.__hasActiveOutsideClickHandling = true;
-    }
-  }
-
-  _checkForOutsideClick(event) {
-    const outsideGroupClicked = !this.contains(event.target);
-    if (outsideGroupClicked) {
-      this.touched = true;
-    }
-  }
-
-  __requestChildrenToBeDisabled() {
-    this.formElementsArray.forEach(child => {
-      if (child.makeRequestToBeDisabled) {
-        child.makeRequestToBeDisabled();
-      }
-    });
-  }
-
-  __retractRequestChildrenToBeDisabled() {
-    this.formElementsArray.forEach(child => {
-      if (child.retractRequestToBeDisabled) {
-        child.retractRequestToBeDisabled();
-      }
-    });
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  inputGroupTemplate() {
-    return html`
-      <div class="input-group">
-        <slot></slot>
-      </div>
-    `;
-  }
-
-  submitGroup() {
-    this.submitted = true;
-    this.formElementsArray.forEach(child => {
-      if (typeof child.submitGroup === 'function') {
-        child.submitGroup();
-      } else {
-        child.submitted = true; // eslint-disable-line no-param-reassign
-      }
-    });
-  }
-
-  serializeGroup() {
-    // Get serializedValue and filter out disabled?
-    // return this._getFromAllFormElements('serializedValue', (node) =>);
-    if (!this.formElements) {
-      return undefined;
-    }
-    const childrenNames = Object.keys(this.formElements);
-    const serializedValues = {};
-    childrenNames.forEach(name => {
-      const element = this.formElements[name];
-      if (Array.isArray(element)) {
-        console.log('__serializeElements');
-        serializedValues[name] = this.__serializeElements(element);
-      } else {
-        console.log('__serializeElement');
-        const serializedValue = this.__serializeElement(element);
-        // if (serializedValue || serializedValue === 0) {
-        serializedValues[name] = serializedValue;
-        // }
-      }
-    });
-    return serializedValues;
-  }
-
-  resetGroup() {
-    this.formElementsArray.forEach(child => {
-      if (typeof child.resetGroup === 'function') {
-        child.resetGroup();
-      } else if (typeof child.reset === 'function') {
-        child.reset();
-      }
-    });
-
-    this.resetInteractionState();
-  }
-
-  resetInteractionState() {
-    // TODO: add submitted prop to InteractionStateMixin
-    this.submitted = false;
-    this.touched = false;
-    this.dirty = false;
-    this.formElementsArray.forEach(formElement => {
-      if (typeof formElement.resetInteractionState === 'function') {
-        formElement.resetInteractionState();
-      }
-    });
+    /** @override from FormRegistrarMixin */
+    this._isFormOrFieldset = true;
   }
 
   /**
+   * @desc Returns all serialized values for formElements.
+   * By default, disabled and Unparseable values are filtered out.
+   * @example
+   * // When we submit a form
+   * formOrFieldset.serializeGroup();
+   * // When we want to 'capture' a full serialized form, so that an interrupted
+   * // session can be stored at a later time. This could either be sent to a server
+   * // or saved in localStorage.
+   * formOrFieldset.serializeGroup({ omitDisabled: false, omitUnparseable: false });
    *
-   * @param {*} property
-   * @param {function} condition
+   * @param {object} [filterOptions] allows to disable default filters
+   * @param {boolean} [filterOptions.omitDisabled=true] when submitting a final result to a
+   * server, disabled values are usually left out.
+   * @returns {object} serialized values, following .formElements structure
    */
-  _getFromAllFormElements(property) {
-    if (!this.formElements) {
-      return undefined;
-    }
-    const childrenNames = Object.keys(this.formElements);
-    const values = {};
-    childrenNames.forEach(name => {
-      if (Array.isArray(this.formElements[name])) {
-        // grouped via myName[]
-        // TODO: doesn't exist anymore -> multiple
-        values[name] = this.formElements[name].map(node => node[property]);
-      } else {
-        // not grouped
-        values[name] = this.formElements[name][property];
-      }
-    });
-    return values;
-  }
-
-  _setValueForAllFormElements(property, value) {
-    this.formElementsArray.forEach(el => {
-      el[property] = value; // eslint-disable-line no-param-reassign
-    });
-  }
-
-  async _setValueMapForAllFormElements(property, values) {
-    if (!this.__readyForRegistration) {
-      await this.registrationReady;
-    }
-
-    if (values && typeof values === 'object') {
-      Object.keys(values).forEach(name => {
-        if (Array.isArray(this.formElements[name])) {
-          this.formElements[name].forEach((el, index) => {
-            el[property] = values[name][index]; // eslint-disable-line no-param-reassign
-          });
-        }
-        this.formElements[name][property] = values[name];
-      });
-    }
-  }
-
-  _anyFormElementHas(property) {
-    return Object.keys(this.formElements).some(name => {
-      if (Array.isArray(this.formElements[name])) {
-        return this.formElements[name].some(el => !!el[property]);
-      }
-      return !!this.formElements[name][property];
-    });
-  }
-
-  _anyFormElementHasFeedbackFor(state) {
-    return Object.keys(this.formElements).some(name => {
-      if (Array.isArray(this.formElements[name])) {
-        return this.formElements[name].some(el => !!el.hasFeedbackFor.includes(state));
-      }
-      return !!this.formElements[name].hasFeedbackFor.includes(state);
-    });
-  }
-
-  _everyFormElementHas(property) {
-    return Object.keys(this.formElements).every(name => {
-      if (Array.isArray(this.formElements[name])) {
-        return this.formElements[name].every(el => !!el[property]);
-      }
-      return !!this.formElements[name][property];
-    });
-  }
-
-  /**
-   * Gets triggered by event 'validate-performed' which enabled us to handle 2 different situations
-   *   - react on modelValue change, which says something about the validity as a whole
-   *       (at least two checkboxes for instance) and nothing about the children's values
-   *   - children validity states have changed, so fieldset needs to update itself based on that
-   */
-  __validate(ev) {
-    if (ev && this.isRegisteredFormElement(ev.target)) {
-      this.validate();
-    }
-  }
-
-  _syncFocused() {
-    this.focused = this._anyFormElementHas('focused');
-  }
-
-  _onFocusOut(ev) {
-    const lastEl = this.formElementsArray[this.formElementsArray.length - 1];
-    if (ev.target === lastEl) {
-      this.touched = true;
-    }
-    this.focused = false;
-  }
-
-  _syncDirty() {
-    this.dirty = this._anyFormElementHas('dirty');
-  }
-
-  _setRole(role) {
-    this.setAttribute('role', role || 'group');
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  __serializeElement(element) {
-    if (!element.disabled) {
-      if (typeof element.serializeGroup === 'function') {
-        return element.serializeGroup();
-      }
-      return element.serializedValue;
-    }
-    return undefined;
-  }
-
-  __serializeElements(elements) {
-    const serializedValues = [];
-    elements.forEach(element => {
-      const serializedValue = this.__serializeElement(element);
-      // if (serializedValue || serializedValue === 0) {
-      serializedValues.push(serializedValue);
-      // }
-    });
-    return serializedValues;
-  }
-
-  /**
-   * Adds the element to an object with the child name as a key
-   * Note: this is different to the default behavior of just beeing an array
-   *
-   * @override
-   */
-  addFormElement(child, indexToInsertAt) {
-    const { name } = child;
-    if (!name) {
-      console.info('Error Node:', child); // eslint-disable-line no-console
-      throw new TypeError('You need to define a name');
-    }
-    if (name === this.name && !this._childrenCanHaveSameName) {
-      console.info('Error Node:', child); // eslint-disable-line no-console
-      throw new TypeError(`You can not have the same name "${name}" as your parent`);
-    }
-
-    if (this.disabled) {
-      // eslint-disable-next-line no-param-reassign
-      child.makeRequestToBeDisabled();
-    }
-
-    if (name.substr(-2) === '[]' || this._childNamesCanBeDuplicate) {
-      if (!Array.isArray(this.formElements[name])) {
-        this.formElements[name] = [];
-      }
-      if (indexToInsertAt > 0) {
-        this.formElements[name].splice(indexToInsertAt, 0, child);
-      } else {
-        this.formElements[name].push(child);
-      }
-    } else if (!this.formElements[name]) {
-      this.formElements[name] = child;
-    } else {
-      console.info('Error Node:', child); // eslint-disable-line no-console
-      throw new TypeError(
-        `Name "${name}" is already registered - if you want an array add [] to the end`,
-      );
-    }
-
-    // This is a way to let the child element (a lion-fieldset or lion-field) know, about its parent
-    // eslint-disable-next-line no-param-reassign
-    child.__parentFormGroup = this;
-
-    // aria-describedby of (nested) children
-    let parent = this;
-    while (parent) {
-      this.constructor._addDescriptionElementIdsToField(
-        child,
-        parent._getAriaDescriptionElements(),
-      );
-      // Also check if the newly added child needs to refer grandparents
-      parent = parent.__parentFormGroup;
-    }
-
-    this.validate();
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  get _childrenCanHaveSameName() {
-    return false;
-  }
-
-  // eslint-disable-next-line class-methods-use-this
-  get _childNamesCanBeDuplicate() {
-    return false;
-  }
-
-  /**
-   * Gathers initial model values of all children. Used
-   * when resetGroup() is called.
-   */
-  get _initialModelValue() {
-    return this._getFromAllFormElements('_initialModelValue');
-  }
-
-  /**
-   * Add aria-describedby to child element(field), so that it points to feedback/help-text of
-   * parent(fieldset)
-   * @param {LionField} field - the child: lion-field/lion-input/lion-textarea
-   * @param {array} descriptionElements  - description elements like feedback and help-text
-   */
-  static _addDescriptionElementIdsToField(field, descriptionElements) {
-    const orderedEls = getAriaElementsInRightDomOrder(descriptionElements, { reverse: true });
-    orderedEls.forEach(el => {
-      if (field.addToAriaDescribedBy) {
-        field.addToAriaDescribedBy(el, { reorder: false });
-      }
-    });
-  }
-
-  removeFormElement(child) {
-    const { name } = child;
-    if (name.substr(-2) === '[]' && this.formElements[name]) {
-      const index = this.formElements[name].indexOf(child);
-      if (index > -1) {
-        this.formElements[name].splice(index, 1);
-      }
-    } else if (this.formElements[name]) {
-      delete this.formElements[name];
-    }
-
-    // TODO: Clean up aria references of elements that were ancestors of child.
-    // For this, it would be better if LionField._ariaDescribedby would be an element array from
-    // which you can delete all elems that are not child.contains(descriptionEl), so that the
-    // resulting array can be serialized into a string of ids.
-
-    this.validate();
+  serializeGroup({ omitDisabled = true } = {}) {
+    const filterCondition = el => !(omitDisabled && el.disabled);
+    return this._getFromAllFormElements('serializedValue', filterCondition);
   }
 }
