+++ packages/form-system/stories/15-features-overview.stories.mdx
@@ -34,16 +34,20 @@ For usage and installation please see the appropriate packages.
       return html`
       <lion-form>
         <form>
-          <lion-input
-            name="first_name"
-            label="First Name"
-            .validators="${[new Required()]}"
-          ></lion-input>
-          <lion-input
-            name="last_name"
-            label="Last Name"
-            .validators="${[new Required()]}"
-          ></lion-input>
+          <lion-fieldset name="full_name">
+            <lion-input
+              name="first_name"
+              label="First Name"
+              .validators="${[new Required()]}"
+            ></lion-input>
+            <lion-input
+              name="last_name"
+              label="Last Name"
+              .validators="${[new Required()]}"
+            ></lion-input>
+            <lion-checkbox name="nsme[]"></lion-checkbox>
+            <lion-checkbox name="nsme[]"></lion-checkbox>
+          </lion-fieldset>
           <lion-input-date
             name="date"
             label="Date of application"
