+++ packages/radio-group/src/LionRadioGroup.js
@@ -1,12 +1,13 @@
+import { LitElement } from '@lion/core';
 import { ChoiceGroupMixin } from '@lion/choice-input';
-import { LionFieldset } from '@lion/fieldset';
+import { FormGroupMixin } from '@lion/fieldset';
 
 /**
  * A wrapper around multiple radios.
  *
  * @extends {LionFieldset}
  */
-export class LionRadioGroup extends ChoiceGroupMixin(LionFieldset) {
+export class LionRadioGroup extends ChoiceGroupMixin(FormGroupMixin(LitElement)) {
   connectedCallback() {
     // eslint-disable-next-line wc/guard-super-call
     super.connectedCallback();
